Add ad_group_membership resource
This resource allows us to assign AD objects such as users and groups
to a given group.

In this PR we are  also adding forceArray to `RunWinRMCommand` because
we need to work around PowerShell's `ConvertTo-Json` command
peculiarity. Specifically if there is only one input object it will
return a hash but if there are many it will return a list. The problem
is that sometimes we expect a list with a single item in it, in which
case we need to wrap the JSON document in `[]` so it looks like a list.