New Functionality - Ability to Specify DC
A very simple addition that allows users to specify a particular domain
controller when desired. Previously, the provider assumed it should
interact with the realm value. Typically this would be fine. But in
cases where several domain controllers are load balanced, there is a
chance an AD object will not synchronize between domain controllers
quickly enough. So, subsequent requests might find unexpected states of
AD objects if the domain controllers have not yet synchronized (AD
is *eventually* consistent). This can be particularly problematic if
domain controllers are physically distant.